+++ src/android/src/main/java/com/reactnativecommunity/slider/ReactSliderDrawable/BackgroundDrawableHandler.java
@@ -1,0 +1,9 @@
+package com.reactnativecommunity.slider.ReactSliderDrawable;
+
+import com.reactnativecommunity.slider.ReactSlider;
+
+public class BackgroundDrawableHandler extends ProgressDrawableHandler {
+  public BackgroundDrawableHandler(ReactSlider slider) {
+    super(slider, android.R.id.background);
+  }
+}
