+++ src/android/src/main/java/com/reactnativecommunity/slider/ReactSliderDrawable/DrawListenerRegistry.java
@@ -1,0 +1,34 @@
+package com.reactnativecommunity.slider.ReactSliderDrawable;
+
+import android.view.View;
+import android.view.ViewTreeObserver;
+
+class DrawListenerRegistry {
+  static void registerListener(final View view, final ViewTreeObserver.OnDrawListener listener) {
+    ViewTreeObserver treeObserver = view.getViewTreeObserver();
+    try {
+      treeObserver.addOnDrawListener(listener);
+    } catch (Throwable throwable) {
+      view.post(new Runnable() {
+        @Override
+        public void run() {
+          registerListener(view, listener);
+        }
+      });
+    }
+  }
+
+  static void unregisterListener(final View view, final ViewTreeObserver.OnDrawListener listener) {
+    ViewTreeObserver treeObserver = view.getViewTreeObserver();
+    try {
+      treeObserver.removeOnDrawListener(listener);
+    } catch (Throwable throwable) {
+      view.post(new Runnable() {
+        @Override
+        public void run() {
+          unregisterListener(view, listener);
+        }
+      });
+    }
+  }
+}
