+++ src/android/src/main/java/com/reactnativecommunity/slider/ReactSliderDrawable/ReactSliderDrawableHelper.java
@@ -1,0 +1,19 @@
+package com.reactnativecommunity.slider.ReactSliderDrawable;
+
+public class ReactSliderDrawableHelper {
+
+  static final int MAX_LEVEL = 10000;
+
+  /*
+  static class ReactDrawableGroup extends LayerDrawable {
+    ReactDrawableGroup(View view) {
+      if (view instanceof ViewGroup) {
+
+      }
+      super();
+    }
+  }
+
+   */
+
+}
